pyproject.toml
[tool.poetry]
name = "maap-api-nasa"
version = "4.3.0a2"
description = "NASA Python implementation of the MAAP API specification"
authors = ["MAAP-Project Platform <platform@MAAP-Project.github.com>"]
license = "Apache 2.0"
readme = "README.md"
packages = [
    {include = "api"}
]
include = [
    { path = "logging.conf", format = ["sdist", "wheel"] }
]

[tool.poetry.dependencies]
python = "^3.9.0"
cachetools = "^5.3.3"
gunicorn = "^22.0.0"
requests = "^2.31.0"
pyshp = "^2.3.1"
gitpython = "^3.1.43"
pyjwt = "^2.8.0"
xmltodict = "^0.13.0"
boto3 = "^1.34.91"
owslib = "^0.30.0"
requests-file = "^2.0.0"
natsort = "^8.4.0"
pillow = "^10.3.0"
pyproj = "^3.6.1"
psycopg2 = {version = "^2.9.9", platform = "linux"}
psycopg2-binary = {version = "^2.9.9", platform = "darwin"}
marshmallow-sqlalchemy = "^1.0.0"
mapproxy = "^2.0.2"
pycryptodome = "^3.20.0"
werkzeug = "^2.2"
flask = "^2"
flask-cors = "^4.0.0"
flask-restx = "^1.3.0"
flask-api = "^3.1"
flask-sqlalchemy = "^3.1.1"
cryptography = "^43.0.0"
cwl-utils = ">0.10"
python-gitlab = "^4.11.1"


[tool.poetry.group.dev.dependencies]
pylint = "^3.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
